add batching configuration and refactor action worker code
