+++ packages/backend/src/config/batches.ts
@@ -1,0 +1,3 @@
+export const QUEUE_BATCH_SIZE = {
+  'm365-excel': 5, // TODO: Update this upon discussions, this should be equal to the group limit concurrency
+}
